renamed component folder [PT #141915635]
