+++ src/modules/sharedComponents/FileUploader/components/Confirmation.js
