+++ src/modules/sharedComponents/FileUploader/components/Confirmation.scss
