+++ src/modules/sharedComponents/FileUploader/components/Confirmation.snapshot.test.js
