+++ src/modules/sharedComponents/FileUploader/components/FileStepper.scss
