+++ src/modules/sharedComponents/FileUploader/components/FileStepper.snapshot.test.js
