+++ src/modules/sharedComponents/FileUploader/components/FileStepper.test.js
