+++ src/modules/sharedComponents/FileUploader/components/FileUploadInfoRow.js
