+++ src/modules/sharedComponents/FileUploader/components/FileUploadInfoRow.scss
