+++ src/modules/sharedComponents/FileUploader/components/FileUploadInfoRow.snapshot.test.js
