+++ src/modules/sharedComponents/FileUploader/components/FileUploader.js
