+++ src/modules/sharedComponents/FileUploader/components/FileUploader.snapshot.test.js
