+++ src/modules/sharedComponents/FileUploader/components/GettingStarted.js
