+++ src/modules/sharedComponents/FileUploader/components/Metadata.snapshot.test.js
