+++ src/modules/sharedComponents/FileUploader/components/UploadDialog.js
