+++ src/modules/sharedComponents/FileUploader/components/UploadDialog.snapshot.test.js
