+++ src/modules/sharedComponents/FileUploader/components/UploadedFileMetadata.js
