+++ src/modules/sharedComponents/FileUploader/components/UploadedFileMetadata.scss
