+++ src/modules/sharedComponents/FileUploader/components/UploadedFileMetadata.snapshot.test.js
