+++ src/modules/sharedComponents/FileUploader/components/fileHelper.js
