+++ src/modules/sharedComponents/FileUploader/containers/Confirmation.js
