+++ src/modules/sharedComponents/FileUploader/containers/GettingStarted.js
