+++ src/modules/sharedComponents/FileUploader/containers/UploadedFileMetadata.js
