+++ src/modules/sharedComponents/FileUploader/reducer.js
